rename validations directory
